use size_t for nnz

+++ main.cpp
@@ -5,7 +5,7 @@
 #include <cstdlib>
 #include "cholmod.h"
 
-cholmod_sparse* generate_sparse_matrix(int rows, int cols, int nnz, cholmod_common* c) {
+cholmod_sparse* generate_sparse_matrix(int rows, int cols, size_t nnz, cholmod_common* c) {
     cholmod_triplet* T = cholmod_allocate_triplet(rows, cols, nnz, 0, CHOLMOD_REAL, c);
     if (!T) {
         std::cerr << "Failed to allocate triplet matrix" << std::endl;
@@ -51,7 +51,7 @@ int main(int argc, char* argv[]) {
         return 1;
     }
 
-    int n = std::atoi(argv[1]);  // Convert argument to integer
+    size_t n = std::atoi(argv[1]);  // Convert argument to integer
     if (n <= 0) {
         std::cerr << "Error: Matrix size must be a positive integer." << std::endl;
         return 1;
@@ -63,7 +63,7 @@ int main(int argc, char* argv[]) {
     std::ofstream csv_file(csv_file_name);
     csv_file << "Matrix_Size,Nonzeros,SpMV_Time(s)\n";
 
-    int nnz = n * n / 100;  // Set sparsity to ~1% nonzero
+    size_t nnz = n * n / 100;  // Set sparsity to ~1% nonzero
     cholmod_sparse* A = generate_sparse_matrix(n, n, nnz, &c);
 
     if (!A) {
